# clacheless-api-rest/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "clacheless_api_rest",
    "description": "Clacheless REST API",
    "license": {
      "name": "Apache-2.0 WITH AdditionRef-FWM-Exception-1.0.0",
      "identifier": "Apache-2.0 WITH AdditionRef-FWM-Exception-1.0.0"
    },
    "version": "0.0.0"
  },
  "paths": {
    "/cache/{cache_key}": {
      "put": {
        "tags": [
          "cache"
        ],
        "summary": "Storing a cached item by key.",
        "operationId": "put_object",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "description": "Cache key.",
            "required": true
          },
          {
            "name": "cache_key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No content. Successfully cached item."
          },
          "400": {
            "description": "Bad Request."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/cache/{key}": {
      "get": {
        "tags": [
          "cache"
        ],
        "summary": "Retrieve a cached item by key.",
        "operationId": "get_object",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "description": "Cache key.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Return the cached object.",
            "content": {
              "application/json": {}
            }
          },
          "404": {
            "description": "No cached item with the key was found."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/health": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "This endpoint returns the combined status of initialized, readiness and\nliveness of a microservice.\n\nCorresponds to the Kubernetes readiness probe.",
        "operationId": "health",
        "responses": {
          "200": {
            "description": "Up",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "HTTP response body object for health requests. Only basic status is\nsupported.",
                  "required": [
                    "status"
                  ],
                  "properties": {
                    "status": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Undetermined"
          },
          "503": {
            "description": "Down"
          }
        }
      }
    },
    "/health/live": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "This endpoint returns the liveness of a microservice, or whether it\nencountered a bug or deadlock. If this check fails, the microservice is not\nrunning and can be stopped.\n\nThis endpoint corresponds to the Kubernetes liveness probe, which automatically\nrestarts the pod if the check fails.",
        "operationId": "health_live",
        "responses": {
          "200": {
            "description": "Up",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "HTTP response body object for health requests. Only basic status is\nsupported.",
                  "required": [
                    "status"
                  ],
                  "properties": {
                    "status": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Undetermined"
          },
          "503": {
            "description": "Down"
          }
        }
      }
    },
    "/health/ready": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "This endpoint returns the readiness of a microservice, or whether it is ready\nto process requests.\n\nCorresponds to the Kubernetes readiness probe.",
        "operationId": "health_ready",
        "responses": {
          "200": {
            "description": "Up",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "HTTP response body object for health requests. Only basic status is\nsupported.",
                  "required": [
                    "status"
                  ],
                  "properties": {
                    "status": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Undetermined"
          },
          "503": {
            "description": "Down"
          }
        }
      }
    },
    "/health/started": {
      "get": {
        "tags": [
          "health"
        ],
        "summary": "In MicroProfile Health 3.1 and later, you can use this endpoint to determine\nwhether your deployed applications are initialized, according to criteria that\nyou define.\n\nCorresponds to the Kubernetes startup probe.",
        "operationId": "health_started",
        "responses": {
          "200": {
            "description": "Up",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "HTTP response body object for health requests. Only basic status is\nsupported.",
                  "required": [
                    "status"
                  ],
                  "properties": {
                    "status": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Undetermined"
          },
          "503": {
            "description": "Down"
          }
        }
      }
    }
  },
  "components": {}
}